test code cleaned
